# Conflicts:
#	src/Cody.Core/Agent/TextDocumentNotificationHandlers.cs
#	src/Cody.VisualStudio/CodyPackage.cs

+++ src/Cody.Core/Agent/TextDocumentNotificationHandlers.cs
@@ -1,11 +1,8 @@
 using Cody.Core.Agent.Protocol;
 using Cody.Core.Common;
 using Cody.Core.Infrastructure;
+using Cody.Core.Logging;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Cody.Core.Agent
 {
@@ -13,25 +10,36 @@ namespace Cody.Core.Agent
     {
         private readonly IDocumentService documentService;
         private readonly IFileDialogService fileDialogService;
+        private readonly ILog logger;
         private readonly IStatusbarService statusbarService;
 
         public TextDocumentNotificationHandlers(IDocumentService documentService,
-            IFileDialogService fileDialogService, IStatusbarService statusbarService)
+            IFileDialogService fileDialogService, IStatusbarService statusbarService, ILog logger)
         {
             this.documentService = documentService;
             this.fileDialogService = fileDialogService;
+            this.logger = logger;
             this.statusbarService = statusbarService;
         }
 
         [AgentCallback("window/showSaveDialog", deserializeToSingleObject: true)]
         public string ShowSaveDialog(SaveDialogOptionsParams paramValues)
         {
-            var filePath = fileDialogService.ShowSaveFileDialog(
-                paramValues.DefaultUri?.ToWindowsPath(), paramValues.Title, paramValues.Filters);
+            try
+            {
+                var filePath = fileDialogService.ShowSaveFileDialog(
+                    paramValues.DefaultUri?.ToWindowsPath(), paramValues.Title, paramValues.Filters);
 
-            var uri = filePath?.ToUri();
+                var uri = filePath?.ToUri();
 
-            return uri;
+                return uri;
+            }
+            catch (Exception ex)
+            {
+                logger.Error("Save file dialog failed.", ex);
+            }
+
+            return null;
         }
 
         [AgentCallback("textDocument/edit", deserializeToSingleObject: true)]
@@ -42,44 +50,73 @@ namespace Cody.Core.Agent
                 var path = textDocumentEdit.Uri.ToWindowsPath();
                 return documentService.EditTextInDocument(path, textDocumentEdit.Edits);
             }
+            catch (Exception ex)
+            {
+                logger.Error($"Document edit failed for '{textDocumentEdit.Uri}'", ex);
+                return false;
+            }
             finally
             {
                 statusbarService.StopProgressAnimation();
             }
+
         }
 
         [AgentCallback("textDocument/show", deserializeToSingleObject: true)]
         public bool ShowTextDocument(TextDocumentShowParams textDocumentShow)
         {
-            var result = documentService.ShowDocument(textDocumentShow.Uri.ToWindowsPath(), textDocumentShow.Options?.Selection);
-            return result;
+            try
+            {
+                var result = documentService.ShowDocument(textDocumentShow.Uri.ToWindowsPath(), textDocumentShow.Options?.Selection);
+                return result;
+            }
+            catch (Exception ex)
+            {
+                logger.Error($"Document show failed for '{textDocumentShow.Uri}'", ex);
+            }
+
+            return false;
         }
 
         [AgentCallback("workspace/edit", deserializeToSingleObject: true)]
         public bool WorkspaceEdit(WorkspaceEditParams workspaceEdit)
         {
+            if (workspaceEdit?.Operations == null)
+            {
+                logger.Error("Workspace edit failed - operations are NULL.");
+                return false;
+            }
+
             try
             {
                 foreach (var operation in workspaceEdit.Operations)
                 {
-                    bool result = false;
-                    switch (operation)
+                    bool result;
+                    try
                     {
-                        case CreateFileOperation createFile:
-                            result = documentService.CreateDocument(createFile.Uri.ToWindowsPath(), createFile.TextContents, createFile.Options?.Overwrite ?? false);
-                            break;
-                        case RenameFileOperation renameFile:
-                            result = documentService.RenameDocument(renameFile.OldUri.ToWindowsPath(), renameFile.NewUri.ToWindowsPath());
-                            break;
-                        case DeleteFileOperation deleteFile:
-                            result = documentService.DeleteDocument(deleteFile.Uri.ToWindowsPath());
-                            break;
-                        case EditFileOperation editFile:
-                            result = documentService.EditTextInDocument(editFile.Uri.ToWindowsPath(), editFile.Edits);
-                            break;
+                        switch (operation)
+                        {
+                            case CreateFileOperation createFile:
+                                result = documentService.CreateDocument(createFile.Uri.ToWindowsPath(), createFile.TextContents, createFile.Options?.Overwrite ?? false);
+                                break;
+                            case RenameFileOperation renameFile:
+                                result = documentService.RenameDocument(renameFile.OldUri.ToWindowsPath(), renameFile.NewUri.ToWindowsPath());
+                                break;
+                            case DeleteFileOperation deleteFile:
+                                result = documentService.DeleteDocument(deleteFile.Uri.ToWindowsPath());
+                                break;
+                            case EditFileOperation editFile:
+                                result = documentService.EditTextInDocument(editFile.Uri.ToWindowsPath(), editFile.Edits);
+                                break;
+                            default:
+                                throw new NotSupportedException($"Not supported operation: {operation}");
+                        }
                     }
-
-                    if (!result) return false;
+                    catch (Exception ex)
+                    {
+                        logger.Error($"Workspace operation ('{operation.Type}') failed for '{GetOperationUri(operation)}'", ex);
+                        return false;
+                    }
                 }
 
                 return true;
@@ -87,6 +124,23 @@ namespace Cody.Core.Agent
             finally
             {
                 statusbarService.StopProgressAnimation();
+            }
+        }
+
+        private string GetOperationUri(WorkspaceEditOperation operation)
+        {
+            switch (operation)
+            {
+                case CreateFileOperation createFile:
+                    return createFile.Uri;
+                case RenameFileOperation renameFile:
+                    return renameFile.OldUri;
+                case DeleteFileOperation deleteFile:
+                    return deleteFile.Uri;
+                case EditFileOperation editFile:
+                    return editFile.Uri;
+                default:
+                    return "unknown";
             }
         }
     }
